Successful refactor into using namespaces.
The error was that the namespaces did not have one top-level namespace. Setting the namespaces as `Core\Database` and `AppSpecific\Controllers` meant that the controllers and the database helper class were prevented from communicating with each other. There has to be an identical top-level namespace for them to use each other. Correct namespaces are `App\Core\Database` and `App\AppSpecific\Controllers`. The keyword `App` is arbitrary.
    - Namespaces are arbitrary and do not need to mimic folder structure to work. Namespaces mimicing folder structure is just a convention.
- Deleted some unnecessary files.